@@ -1,54 +1,62 @@
 {
-  "name": "wedemyclient",
-  "version": "1.2.0",
+  "name": "wedemyvite",
+  "version": "2.0.0",
   "private": true,
   "author": "Davis Tibbz",
   "license": "MIT",
-  "homepage": "https://github.com/Longwater1234/WedemyClient",
+  "homepage": "https://github.com/Longwater1234/wedemyvite",
   "engines": {
-    "node": ">=16.x"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Longwater1234/WedemyClient.git"
-  },
-  "bugs": {
-    "url": "https://github.com/Longwater1234/WedemyClient/issues"
+    "node": ">= 16.0"
   },
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "prod": "npx http-server ./dist --proxy http://localhost:8080?"
+    "dev": "cross-env --max_old_space_size=4096 vite",
+    "build": "rimraf dist && vite build",
+    "preview": "vite preview",
+    "type-check": "vue-tsc --noEmit",
+    "lint": "eslint src --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix",
+    "format": "prettier --write src/**",
+    "prepare": "husky install"
   },
   "dependencies": {
-    "@element-plus/icons-vue": "^1.1.4",
-    "@hcaptcha/vue3-hcaptcha": "^1.0.1",
+    "@element-plus/icons-vue": "^2.1.0",
     "@techassi/vue-youtube-iframe": "^1.0.6",
-    "axios": "^0.26.1",
-    "braintree-web-drop-in": "^1.6.1",
-    "core-js": "^3.8.3",
-    "element-plus": "^2.1.8",
-    "prettier": "^2.7.1",
-    "register-service-worker": "^1.7.2",
-    "validator": "^13.7.0",
-    "vue": "^3.2.13",
-    "vue-github-button": "^3.0.3",
-    "vue-router": "^4.0.3",
-    "vue2-timeago": "^3.0.2"
+    "@vueuse/core": "^9.13.0",
+    "axios": "^1.3.5",
+    "braintree-web-drop-in": "^1.36.1",
+    "element-plus": "^2.3.3",
+    "pinia": "^2.0.36",
+    "vue": "^3.2.47",
+    "vue-eslint-parser": "^9.1.1",
+    "vue-github-button": "^3.1.0",
+    "vue-router": "^4.2.0",
+    "vue-timeago3": "^2.3.0"
   },
   "devDependencies": {
     "@hcaptcha/types": "^1.0.3",
-    "@types/braintree-web-drop-in": "^1.22.4",
+    "@rushstack/eslint-patch": "^1.2.0",
+    "@types/braintree-web-drop-in": "^1.34.1",
+    "@types/node": "^18.14.2",
+    "@types/validator": "^13.7.15",
     "@types/youtube": "^0.0.47",
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-plugin-pwa": "~5.0.0",
-    "@vue/cli-plugin-router": "~5.0.0",
-    "@vue/cli-plugin-typescript": "~5.0.0",
-    "@vue/cli-service": "~5.0.0",
-    "@vue/reactivity": "^3.2.45",
-    "css-loader": "^6.7.1",
-    "sass": "^1.32.7",
-    "sass-loader": "^12.0.0",
-    "typescript": "~4.5.5"
+    "@vitejs/plugin-vue": "^4.0.0",
+    "@vitejs/plugin-vue-jsx": "^3.0.0",
+    "@vue/eslint-config-prettier": "^7.1.0",
+    "@vue/eslint-config-typescript": "^11.0.2",
+    "@vue/tsconfig": "^0.1.3",
+    "cross-env": "^7.0.3",
+    "eslint": "^8.34.0",
+    "eslint-plugin-unused-imports": "^2.0.0",
+    "eslint-plugin-vue": "^9.9.0",
+    "husky": "^8.0.0",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.8.8",
+    "pretty-quick": "^3.1.3",
+    "rimraf": "^3.0.2",
+    "sass": "^1.62.0",
+    "typescript": "^4.8.4",
+    "vite": "^4.3.6",
+    "vite-plugin-eslint": "^1.8.1",
+    "vite-plugin-pwa": "^0.14.7",
+    "vue-tsc": "^1.6.5"
   }
 }
