@@ -3,7 +3,7 @@
   <div class="mobile-only">
     <el-card shadow="never" class="m-card">
       <p>Accepting all major credit cards and PayPal.</p>
-      <h2>Total: ${{ totalPrice }}</h2>
+      <h2>Total: ${{ props.totalPrice }}</h2>
       <div>
         <router-link to="/checkout">
           <el-button id="checkout-btn" class="btn purple">
@@ -15,30 +15,21 @@
   </div>
 </template>
 
-<script lang="ts">
-import { defineComponent } from "vue";
+<script lang="ts" setup>
 import { ArrowRight } from "@element-plus/icons-vue";
 
-export default defineComponent({
-  name: "MobileSummary",
-  data() {
-    return {};
-  },
-  props: {
-    totalPrice: {
-      type: String,
-      default: "0.0",
-    },
-  },
-  components: {
-    ArrowRight,
-  },
+const props = defineProps({
+  totalPrice: {
+    type: Number,
+    required: true,
+    default: 0.0
+  }
 });
 </script>
 
 <style scoped>
 .m-card {
-  width: 96vw;
+  width: 100vw;
   font-size: small;
   margin-bottom: 1em;
 }
