@@ -1,168 +1,34 @@
 <template>
-  <Navbar :categories="categories" v-if="!this.$route.meta.hideNavbar" />
+  <Navbar :categories="categories" v-if="!computedHideNavbar" />
   <router-view />
-  <Footer v-if="!this.$route.meta.hideFooter" />
+  <Footer v-if="!computedHideFooter" />
 </template>
 
-<script lang="ts">
+<script lang="ts" setup>
 import Navbar from "@/components/Navbar.vue";
 import Footer from "@/components/Footer.vue";
-import store from "@/store";
-import { defineComponent } from "@vue/runtime-core";
-import CourseService from "./services/CourseService";
 
-export default defineComponent({
-  name: "App",
-  components: {
-    Navbar,
-    Footer,
-  },
-  data() {
-    return {
-      categories: new Array<{ id: number; category: string }>(),
-    };
-  },
-  methods: {
-    fetchCartCount() {
-      store.getCartCountServer().then(() => {});
-    },
-  },
-  mounted() {
-    store.getAuthStatusServer().then((loggedIn) => {
-      if (loggedIn) this.fetchCartCount();
-    });
-    CourseService.getAllCategories().then(
-      (res) => (this.categories = res.data)
-    );
-  },
+import CourseService from "@/service/CourseService";
+import { useStudentStore } from "@/stores";
+import { computed, onMounted, ref } from "vue";
+import type { CategoryDto } from "@/interfaces/custom";
+import { useRoute } from "vue-router";
+
+const store = useStudentStore();
+const categories = ref<CategoryDto[]>([]);
+const route = useRoute();
+
+function fetchCartCount() {
+  store.getCartCountServer().then(() => {});
+}
+
+onMounted(() => {
+  store.getLoginStatus().then((loggedIn: boolean) => {
+    if (loggedIn) fetchCartCount();
+  });
+  CourseService.getAllCategories().then(res => (categories.value = res.data));
 });
+
+const computedHideNavbar = computed(() => route.meta?.hideNavbar ?? false);
+const computedHideFooter = computed(() => route.meta?.hideFooter ?? false);
 </script>
-
-<style>
-* {
-  font-family: "Fira Sans", system-ui, sans-serif;
-}
-:root {
-  --primary: #5943be;
-  --secondary: #3b2b83;
-  --background: #ffff;
-  --el-rate-text-color: var(--background);
-  --dark: #1c1d1f;
-}
-#app {
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  background-color: var(--background);
-  max-height: 100vh;
-  width: auto;
-}
-
-.btn {
-  width: 100%;
-  height: 3em !important;
-}
-
-.btn.purple {
-  background-color: var(--primary);
-  font-weight: 700;
-  color: white;
-}
-
-.btn.purple:hover {
-  background-color: var(--secondary);
-  color: white;
-}
-
-.btn.white {
-  background-color: aliceblue;
-  font-weight: 700;
-  color: var(--primary);
-}
-
-.main-view {
-  padding: 0 10%;
-}
-
-.server-error {
-  color: red;
-}
-
-a {
-  text-decoration: none;
-}
-
-.product-img {
-  aspect-ratio: 16/9;
-  height: 10em;
-  transition: ease-in-out 0.6s;
-}
-.product-img:hover {
-  transform: scale(1.2);
-}
-
-.card-title {
-  font-size: 20px;
-  font-weight: 700;
-}
-.card-author {
-  color: #636362;
-  font-size: 14px;
-}
-
-.nodata {
-  display: flex;
-  margin: 3em auto;
-  flex-direction: row;
-  justify-content: center;
-  color: grey;
-}
-
-.cart-header {
-  font-family: Georgia, "Times New Roman", Times, serif;
-  background-color: black;
-  color: white;
-  margin-top: 0;
-  height: 2em;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  top: 0;
-  padding: 2% 10%;
-}
-
-.noblue {
-  color: black !important;
-}
-
-.mobile-only {
-  display: none;
-}
-
-@media only screen and (max-width: 700px) {
-  .product-img {
-    aspect-ratio: 16/9;
-    height: 10em;
-  }
-
-  .main-body {
-    width: inherit !important;
-  }
-
-  .full-only {
-    display: none;
-  }
-  .cart-header {
-    padding: 5% 10%;
-    height: 2em;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-  }
-
-  .mobile-only {
-    display: unset;
-    height: auto;
-    max-width: 100%;
-  }
-}
-</style>
