@@ -3,14 +3,8 @@
     <div @click="launch = true" class="drawer">
       <img src="@/assets/menu.png" alt="More" class="hamburger" />
     </div>
-    <el-drawer
-      v-model="launch"
-      direction="ltr"
-      size="60%"
-      custom-class="demo-drawer"
-      ref="drawer"
-    >
-      <div class="auth-buttons" v-if="!store.getters.isLoggedIn">
+    <el-drawer v-model="launch" direction="ltr" size="60%" class="demo-drawer" ref="drawer">
+      <div class="auth-buttons" v-if="!store.getIsLoggedIn">
         <router-link to="/login">
           <el-button class="btn purple">Log in</el-button>
         </router-link>
@@ -21,15 +15,10 @@
       </div>
 
       <div style="margin: 1em" v-else>
-        <el-avatar :src="attachAvatarLink(store.state.username)" :size="36">
-        </el-avatar>
-        <span style="margin-left: 10px">{{ store.state.username }}</span>
+        <el-avatar :src="attachAvatarLink(store.fullname)" :size="36"> </el-avatar>
+        <span style="margin-left: 10px">{{ store.fullname }}</span>
         <div style="margin-top: 2em">
-          <el-row
-            v-for="item in navMenuList"
-            :key="item.id"
-            @click="goTo(item.url)"
-          >
+          <el-row v-for="item in navMenuList" :key="item.id" @click="router.push(item.url)">
             {{ item.title }}
             <el-divider />
           </el-row>
@@ -40,38 +29,29 @@
   </div>
 </template>
 
-<script lang="ts">
-import AuthService from "@/services/AuthService";
-import store from "@/store";
-import { defineComponent } from "vue";
-import navMenuList from "@/navmenu.json";
+<script lang="ts" setup>
+import { ref } from "vue";
+import navMenuList from "../navmenu.json";
+import { useStudentStore } from "@/stores";
+import { useRouter } from "vue-router";
 
-export default defineComponent({
-  name: "Drawer",
-  inject: ["store"],
-  data() {
-    return {
-      launch: false,
-      navMenuList,
-    };
-  },
-  methods: {
-    attachAvatarLink: (username: string) => {
-      return `https://avatars.dicebear.com/api/initials/${username}.svg`;
-    },
-    logout: async () => {
-      await AuthService.logoutUser();
-      await store.getAuthStatusServer();
-      window.location.replace("/");
-    },
-    goTo(url: string) {
-      this.$router.push(url);
-    },
-  },
-});
+const store = useStudentStore();
+
+const launch = ref(false);
+const router = useRouter();
+
+const attachAvatarLink = (username: string) => {
+  return `https://avatars.dicebear.com/api/initials/${username}.svg`;
+};
+
+const logout = async () => {
+  await store.logoutUser();
+  await store.getLoginStatus();
+  window.location.replace("/");
+};
 </script>
 
-<style>
+<style scoped>
 .drawer {
   width: 30px;
   height: 30px;
@@ -86,10 +66,7 @@ export default defineComponent({
   width: 100%;
   height: 100%;
 }
-.sidename {
-  color: rgb(61, 61, 61);
-  margin: 0 10px;
-}
+
 .auth-buttons {
   width: 70%;
   margin: 0 auto;
