@@ -1,11 +1,10 @@
-<!-- LOGIN -->
 <!-- Copyright (c) 2022. Davis Tibbz. Github: https://github.com/longwater1234. MIT License  -->
 <template>
   <div align="center" style="height: 80vh">
     <div class="loginContainer">
       <h3 class="loginHeader">Login to your Wedemy Account</h3>
 
-      <!-- GOOGLE SIGN IN  -->
+      <!-- GOOGLE SIGN IN, SEE DOCS  -->
       <!-- https://developers.google.com/identity/gsi/web/guides/display-button -->
       <div
         id="g_id_onload"
@@ -28,16 +27,10 @@
       <!-- END OF GOOGLE BUTTON -->
 
       <!-- START LOGIN FORM BELOW -->
-      <el-form
-        @submit.prevent
-        status-icon
-        :model="loginForm"
-        :rules="rules"
-        ref="loginForm"
-      >
+      <el-form @submit.prevent="handleLogin" status-icon :model="loginForm" :rules="rules" ref="loginFormRef">
         <el-form-item style="margin-top: 10px" prop="email">
           <el-input
-            native-type="email"
+            type="email"
             :prefix-icon="Message"
             placeholder="E-mail"
             maxlength="70"
@@ -53,7 +46,6 @@
             :prefix-icon="Lock"
             v-model.trim="loginForm.password"
             class="field"
-            native-type="password"
             show-password
           ></el-input>
         </el-form-item>
@@ -68,13 +60,7 @@
         </el-form-item>-->
 
         <div style="margin-top: 8px">
-          <el-button
-            class="btn purple"
-            @click="handleLogin('loginForm')"
-            style="font-weight: bold"
-            native-type="submit"
-            :loading="isLoading"
-          >
+          <el-button class="btn purple" style="font-weight: bold" native-type="submit" :loading="isLoading">
             Log in
           </el-button>
         </div>
@@ -83,135 +69,127 @@
 
       <div style="margin-top: 13px">
         New user?
-        <router-link to="/signup" class="none" :style="{ fontWeight: '800' }">
-          Create an Account
-        </router-link>
+        <router-link to="/signup" class="none" :style="{ fontWeight: '800' }"> Create an Account </router-link>
       </div>
     </div>
   </div>
 </template>
 
-<script>
-import AuthService from "@/services/AuthService";
-import store from "@/store";
+<script lang="ts" setup>
+import AuthService from "@/service/AuthService";
 import { ElMessage } from "element-plus";
-import isEmail from "validator/lib/isEmail";
-import VueHcaptcha from "@hcaptcha/vue3-hcaptcha";
-import { Lock, Message } from "@element-plus/icons-vue/dist/lib";
-import { markRaw } from "@vue/reactivity";
+import { Lock, Message } from "@element-plus/icons-vue";
+import { computed, onBeforeUnmount, onMounted, reactive, ref } from "vue";
+import { handleApiError } from "@/util/http_util";
+import type { FormInstance, FormRules } from "element-plus";
+import { useStudentStore } from "@/stores";
+import type { LoginRequest, UserDto } from "@/interfaces/custom";
 
-export default {
-  name: "Login",
-  data() {
-    document.title = "Login | Wedemy";
-    // validation for email
-    const checkEmail = (rule, value, callback) => {
-      if (!value) {
-        return callback(new Error("E-mail can't be empty"));
-      } else if (!isEmail(this.loginForm.email)) {
-        callback(new Error("Email is invalid"));
-      } else {
-        callback();
-      }
-    };
+document.title = "Login | Wedemy";
 
-    // validation for password
-    const checkPassword = (rule, value, callback) => {
-      if (!value) {
-        callback(new Error("Password can't be empty"));
-      } else {
-        callback();
-      }
-    };
+const loginFormRef = ref<FormInstance>();
+const store = useStudentStore();
 
-    return {
-      loginForm: {
-        email: "",
-        password: "",
-        responseToken: "",
-      },
-
-      // rules for the validation
-      rules: {
-        email: [{ validator: checkEmail, trigger: "blur" }],
-        password: [{ validator: checkPassword, trigger: "blur" }],
-      },
-
-      //other data
-      Message: markRaw(Message),
-      Lock: markRaw(Lock),
-      isLoading: false,
-      GOOGLE_CLIENT_ID: process.env.VUE_APP_GOOGLE_CLIENT_ID,
-      SERVER_ROOT: process.env.VUE_APP_BACKEND_ROOT_URL,
-      //HCAPTCHA_KEY: process.env.VUE_APP_HCAPTCHA_CLIENT_KEY,
-    };
-  },
-  methods: {
-    handleLogin(formName) {
-      this.$refs[formName].validate((valid) => {
-        if (!valid) return;
-        this.isLoading = true;
-        this.submitToServer(this.loginForm)
-          .then(() => this.redirectToHome())
-          .catch((error) => this.displayError(error))
-          .finally(() => (this.isLoading = false));
-      });
-    },
-    async submitToServer(payload) {
-      await AuthService.loginUser({ ...payload });
-      await store.getAuthStatusServer();
-      await store.getCartCountServer();
-    },
-    redirectToHome() {
-      ElMessage.success("Welcome back!");
-      this.$router.replace("/");
-    },
-
-    /** onSuccess captcha solve */
-    handleVerify(token) {
-      this.loginForm.responseToken = token;
-    },
-    displayError(error) {
-      let mama = error.response ? "Wrong credentials!" : error.message;
-      console.error("loginError", error);
-      ElMessage.error(mama);
-      setTimeout(() => {
-        this.resetCaptcha();
-      }, 200);
-    },
-    resetCaptcha() {
-      this.loginForm.responseToken = "";
-      this.$refs.mycaptcha.reset();
-    },
-  },
-  mounted() {
-    //attach GoogleAuth script
-    const scripta = document.createElement("script");
-    scripta.src = `https://accounts.google.com/gsi/client`;
-    scripta.id = "google_client";
-    document.getElementById("baba")?.appendChild(scripta); // see index.html
-  },
-  components: {
-    VueHcaptcha,
-  },
-  beforeUnmount() {
-    //detach above script
-    document.getElementById("google_client")?.remove();
-  },
+// validation for password
+const checkPassword = (rule: any, value: string, callback: (arg?: Error) => void) => {
+  if (!value) {
+    callback(new Error("Password can't be empty"));
+  } else {
+    callback();
+  }
 };
+
+const loginForm = reactive({
+  email: "",
+  password: "",
+  responseToken: ""
+});
+
+// rules for the validation
+const rules = reactive<FormRules>({
+  email: [{ required: true, type: "email", trigger: "blur" }],
+  password: [{ validator: checkPassword, trigger: "blur" }]
+});
+
+const isLoading = ref(false);
+const GOOGLE_CLIENT_ID = computed(() => {
+  return import.meta.env.VITE_APP_GOOGLE_CLIENT_ID;
+});
+const SERVER_ROOT = computed(() => {
+  return import.meta.env.VITE_APP_BACKEND_ROOT_URL;
+});
+//const HCAPTCHA_KEY = import.meta.env.VITE_APP_HCAPTCHA_CLIENT_KEY,
+
+/** on formSubmit */
+async function handleLogin() {
+  const isValid = await loginFormRef.value?.validate();
+  if (!isValid) return;
+  isLoading.value = true;
+  submitToServer(loginForm)
+    .then(() => redirectToHome())
+    .catch(error => displayError(error))
+    .finally(() => (isLoading.value = false));
+}
+
+async function submitToServer(payload: LoginRequest) {
+  let res = await AuthService.loginUser({ ...payload });
+  let user: UserDto = res.data.userInfo;
+  store.$patch({
+    id: user.id,
+    fullname: user.fullname,
+    loggedIn: true
+  });
+  await store.getLoginStatus();
+  await store.getCartCountServer();
+}
+
+function redirectToHome() {
+  ElMessage.success("Welcome back!");
+  window.location.replace("/");
+}
+
+/** onSuccess captcha solve */
+// function handleVerify(token: string) {
+//   loginForm.responseToken = token;
+// }
+
+function displayError(err: any) {
+  handleApiError(err);
+  setTimeout(() => {
+    resetCaptcha();
+  }, 200);
+}
+
+function resetCaptcha() {
+  loginForm.responseToken = "";
+  //this.$refs.mycaptcha.reset();
+}
+
+onMounted(() => {
+  //attach GoogleAuth script
+  const script = document.createElement("script");
+  script.src = "https://accounts.google.com/gsi/client";
+  script.id = "google_client";
+  document.getElementById("g-login")?.appendChild(script); // see index.html
+});
+
+onBeforeUnmount(() => {
+  //detach above script
+  document.getElementById("google_client")?.remove();
+});
 </script>
 
-<style>
+<style scoped>
 .loginHeader {
   border-bottom: solid 1px #d1d7dc;
-  color: #1c1d1f;
+  color: var(--color-text);
   display: block;
   font-weight: 700;
   padding: 24px 64px 24px 24px;
 }
 
 .loginContainer {
-  color: #1c1d1f;
+  color: var(--color-text);
   width: 380px;
   text-align: center;
 }
