@@ -2,13 +2,7 @@
 <template>
   <h3 class="cart-header">My Purchase History</h3>
   <div class="main-view" style="height: 70vh" v-loading="isLoading">
-    <el-alert
-      v-if="serverError"
-      :title="serverError"
-      type="error"
-      :closable="false"
-    >
-    </el-alert>
+    <el-alert v-if="serverError" :title="serverError" type="error" :closable="false" />
 
     <div class="mobile-only">
       <small>Scroll horizontally to see all columns</small>
@@ -25,9 +19,7 @@
       </div>
       <!-- dialog footer -->
       <template #footer>
-        <el-button type="primary" @click="dialogVisible = false">
-          Done
-        </el-button>
+        <el-button type="primary" @click="dialogVisible = false"> Done </el-button>
       </template>
     </el-dialog>
     <!-- END OF DIALOG -->
@@ -36,39 +28,15 @@
 
     <div class="salesList" v-if="sales.length > 0">
       <el-table id="myTable" :data="sales" stripe style="width: 100%">
-        <el-table-column
-          prop="transactionId"
-          label="Transcation ID"
-          width="180"
-        />
-        <el-table-column
-          prop="createdAt"
-          label="Date"
-          :formatter="formatter"
-          width="180"
-        />
-        <el-table-column
-          prop="totalPaid"
-          min-width="150px"
-          label="Total Paid (USD)"
-        />
-        <el-table-column
-          prop="paymentMethod"
-          min-width="150px"
-          label="Payment Method"
-        />
-        <el-table-column
-          prop="numOfItems"
-          min-width="150px"
-          label="Items Bought"
-        />
+        <el-table-column prop="transactionId" label="Transcation ID" width="180" />
+        <el-table-column prop="createdAt" label="Date" :formatter="formatter" width="180" />
+        <el-table-column prop="totalPaid" min-width="150px" label="Total Paid (USD)" />
+        <el-table-column prop="paymentMethod" min-width="150px" label="Payment Method" />
+        <el-table-column prop="numOfItems" min-width="150px" label="Items Bought" />
         <el-table-column min-width="100px" label="">
           <template #default="scope">
             <!-- VIEW DETAILS -->
-            <el-button
-              size="small"
-              @click="handleView(scope.row.transactionId)"
-            >
+            <el-button size="small" @click="handleView(scope.row.transactionId)">
               <info-filled style="width: 1em" /> Details
             </el-button>
           </template>
@@ -81,71 +49,50 @@
     <div v-else class="empty-view">
       <el-empty description="You haven't bought anything yet!"></el-empty>
       <router-link to="/">
-        <el-button plain class="btn purple" style="width: 15em">
-          Keep shopping
-        </el-button>
+        <el-button plain class="btn purple" style="width: 15em"> Keep shopping </el-button>
       </router-link>
     </div>
   </div>
 </template>
 
-<script lang="ts">
-import EnrollService from "@/services/EnrollService";
-import { Course, Sale } from "@/types";
+<script lang="ts" setup>
+import EnrollService from "@/service/EnrollService";
+import type { Course, Sale } from "@/interfaces/wedemy";
 import { InfoFilled } from "@element-plus/icons-vue";
-import { defineComponent, onMounted, ref } from "vue";
+import { onMounted, ref } from "vue";
 
-export default defineComponent({
-  setup() {
-    document.title = "Purchase History | Wedemy";
+const isLoading = ref(true);
+const dialogVisible = ref(false);
+const sales = ref<Sale[]>([]);
+const courseList = ref<Course[]>([]);
+const serverError = ref("");
+const courseLoading = ref(false);
 
-    let isLoading = ref(true);
-    let dialogVisible = ref(false);
-    let sales = ref<Sale[]>([]);
-    let courseList = ref<Course[]>([]);
-    let serverError = ref("");
-    let courseLoading = ref(false);
+const fetchMyPurchaseHistory = () => {
+  EnrollService.getPurchaseHistory()
+    .then(res => (sales.value = res.data.content))
+    .catch(err => (serverError.value = err.message))
+    .finally(() => (isLoading.value = false));
+};
 
-    const fetchMyPurchaseHistory = () => {
-      EnrollService.getPurchaseHistory()
-        .then((res) => (sales.value = res.data.content))
-        .catch((err) => (serverError.value = err.message))
-        .finally(() => (isLoading.value = false));
-    };
+/** format date */
+const formatter = (row: Sale) => {
+  return new Date(row.createdAt).toDateString();
+};
 
-    /** format date */
-    const formatter = (row: Sale) => {
-      return new Date(row.createdAt).toDateString();
-    };
+/** fetch specific transaction by ID */
+const handleView = (id: string) => {
+  dialogVisible.value = true;
+  courseLoading.value = true;
+  EnrollService.getItemsByTransactionId(id)
+    .then(res => (courseList.value = res.data.content))
+    .catch(err => (serverError.value = err.message))
+    .finally(() => (courseLoading.value = false));
+};
 
-    /** fetch specific transaction by ID */
-    const handleView = (id: string) => {
-      dialogVisible.value = true;
-      courseLoading.value = true;
-      EnrollService.getItemsByTransactionId(id)
-        .then((res) => (courseList.value = res.data.content))
-        .catch((err) => (serverError.value = err.message))
-        .finally(() => (courseLoading.value = false));
-    };
-
-    onMounted(() => {
-      fetchMyPurchaseHistory();
-    });
-
-    return {
-      sales,
-      courseList,
-      isLoading,
-      courseLoading,
-      dialogVisible,
-      serverError,
-      formatter,
-      handleView,
-    };
-  },
-  components: {
-    InfoFilled,
-  },
+onMounted(() => {
+  document.title = "Purchase History | Wedemy";
+  fetchMyPurchaseHistory();
 });
 </script>
 
